requirements.txt
imap-tools
python-dotenv
requests
schedule
